feat(tasks): decouple job status check from trigger run (#65)
closes #62
